Reran synthesis with new TCL script.

+++ workloads/workload_PWM/gen_pwm.c
@@ -1,7 +1,7 @@
 #include "gen_pwm.h"
 #include <stdio.h>
 //#include <stdint.h>
-void gen_pwm(int32_t duty, int32_t freq, int32_t * out )
+void gen_pwm(int32_t duty, int32_t freq, int8_t * out )
 {
     int32_t i=0;
     for(i=0;i<freq;i++)
