+++ workloads/workload_PWM/gen_pwm.h
@@ -1,3 +1,3 @@
 //#include <ap_cint.h>
 #include <stdint.h>
-void gen_pwm(int32_t duty, int32_t freq, int32_t * out );
+void gen_pwm(int32_t duty, int32_t freq, int8_t * out );
